requirements.txt
aiogram==3.3.0
selenium==4.20.0
webdriver-manager==4.0.1
aiohttp==3.9.5
beautifulsoup4==4.12.3
python-dotenv==1.0.1
